refac: 型の切り出し、コンポーネントフォルダの配置
