+++ app/view/components/recordItem.tsx
